refactor ALB010 alignment nesting validation logic
The feature file was modified with a custom step '.IfcAlignment. must be paried with .IfcReferent.'.

This step required that the IfcAlignmentHorizontal and the IfcReferent be contained within RelatedObjects of the same nesting relationship.

However, it is possible that the IfcAlignmentHorizontal is nested to IfcAlignment with IfcAlignmentVertical and also (possibly) IfcAlignmentCant with one IfcRelNests, while another instance of IfcRelNests provides the nesting relationship of IfcAlignment as the "whole" and IfcReferent as the "parts".

The revised feature file allows for both options, and re-uses an existing step - which is preferable to introducing a new step and implementation.

(IVS-388)