IVS-75 ALB002 reactivation (#401)
* reactivate and refactor ALB002 feature file

(IVS-75)

* fix unit test file header content

(IVS-75)

* delete existing unit test suite to avoid schema errors

(IVS-75)

* scenarios 01 and 02 passing and failing as expected

(IVS-75)

* adjust ALB003 and SYS001 features to include control characters

Refactoring of step implementations in this manner does not constitute a change in rule version.
All models that previously passed or failed will continue to generate the same outcomes as before.
The control characters provide much more robust parsing of the validation logic in the gherkin statements.

* add fail files for scenario04

(IVS-75)

* add fail files for scenario03

(IVS-75)

* clean up schema errors for duplicate GlobalId values

(IVS-75)

* address review comments

We have an opportunity to re-use 'standard' steps,
versus the bespoke step implementation previously being used.
However, this change causes two the 'pass' files to now fail.
These two files are the ones illustrating the correct re-use of horizontal layouts, which is the main objective of reactivating ALB002.

(IVS-75)

* Remove scenario 2

* Remove nesting step_impl

---------

Co-authored-by: Geert Hesselink <geert.hess@gmail.com>