refactor step implementations to individual functions
The step implementation previously included all matching step permutations in the decorator but also repeated fragments with control flow statements in the body of the function.

This revised functions are much smaller and each match a single step statement.
There are a few function calls repeated with each implementation, but these are cached with `functools` to minimize any performance ramifications.

(IVS-87)